requirements-deployment.txt
fiftyone
valohai-utils
tensorflow
Pillow
python-multipart